refactor(#8/user): 서비스 패키지명을 application 으로 변경
