+++ src/main/java/org/sopt/user/presentation/UserController.java
@@ -7,7 +7,7 @@ import java.net.URI;
 import lombok.RequiredArgsConstructor;
 import org.sopt.global.response.ApiResponse;
 import org.sopt.user.presentation.dto.request.CreateUserRequest;
-import org.sopt.user.service.UserService;
+import org.sopt.user.application.UserService;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
